Support running tests in just one browser.
Add ”test:unit” and ”test:browser” npm scripts, to be able to run unit tests (not launching a browser) and browser tests separately. Additionally check the BROWSERTIME_TEST_BROWSER env variable when running browser tests. If set, it will limit what browsers are used in tests.

This is in preparation for using Travis test matrix for reducing total test time.