Move video out of pre/post scripts
When we first started with the video we used the pre/post structure.
That was ok to move fast but one of the negatives is that stopping the
video happen after we collected all metrics.

Let us move video to separate functionality. This will mostly just change
the structure so the order is correct, then is following PRs
we can tune and move things around and make the code easier to understand
(and change).