Store extra JSON and screenshots per run (and collect stats between runs)
We want to make Browsertime as mean and clean as possible: Store all
extra JSONs (chrome trace categories, console log and more), and
the screenshots between runs (before they where stored on exit).
This is good because it will decrease the memory impact but it is a
non backward compatible change! Sitespeed.io and other tools need to
change how they handle extra JSONs and the screenshot.

I've also moved most stats to be collected between runs, that is
needed for CPU stats since we store the data and throws it away
between runs.