Android tweaks (#512)
* Migrate android code to not use promise chains.

Extract common code into private methods, and update to use async/await instead of promise chains. Also simplify the Android class to mutate its’ state in fewer places.

* Refactor Android code to allow additional exports.

More android specific logic can be extracted that’s not a stateful object. This allows for that.

* Extract detection of Android configuration.

Refactor to detection of Android configuration is done in one place. Update so any Android option is enough to trigger Android, not just ”chrome.android.package”. This allows for ”chrome.android.deviceSerial” to be set, but not ”chrome.android.package” (to fall back to the default Chrome browser).

* Add FIXME reminder about hard coded path.

The Android class resolves the path to external storage by checking an environmental variable on the phone. However the chrome setup assumes it’s at /sdcard.