Remove Firefox version check for HAR export trigger (#647)
We had a check that FF is 61 or later for the HAR export trigger to work. However that fix makes sitespeed.io break when you run with mobile options, see sitespeedio/sitespeed.io#2178 (comment)

It also makes Firefox tests to break as long as you set a different user agent.

We are now on FF 62 so we can just remove the check.