Use orange layer on Android too.
The old implementation use an orange page that was unloaded on unload
on Android. That worked fine but broke when you used a script clicking
on links and measurering visual metrics.

An alternative could be to have the orange page when you don't use a script
but we can fix that later on.