Adjust visual progress timestamps
Currently the timestamps of visual progress are relative to the time of
the first frame of recording. However, in order for visual progress to
be displayed correctly in the Firefox Profiler, we must adjust these
timestamps to be relative to the profiler start time. In this commit, we
adjusted the timestamp of visual progress by adding the offset to
profiler start time, as well as time from first frame to orange frame
which is not captured originally.