Try to find partly orange frames.
We are comparing 40000 pixels and if diff is less than 100 pixels
it is the same. Or was before our change. That doesn't seems correct
since we are trying to find vsync issues with Chrome where parts of
the screen is orange. It should be if the diff < 40000 it is still orange,
but lets be conservative with the change.