Fix font usage on macOS
On macOS, addTextToVideo and getTimingMetrics will fail because the code
makes the assumption /System/Library/Fonts/SFNSText.ttf always exists.

In the new getFont() api, we check that assumption and fallback to
not specifying a font. Also, we make sure to use the font passed
through options.videoParams.fontPath even on macOS because why not.