Add a temporary workaround for orange frame removal issue
Visual metrics are not correct when running browsertime
with the new Firefox for Android browser (Fenix) when
WebRender is enabled, because there's a bug in WebRender.

This is tracking bug in bugzilla:
 - https://bugzilla.mozilla.org/show_bug.cgi?id=1606365

In an ideal world, we fix the bug in WebRender, however I
don't know when that will happen, so this is the workaround.