Add scripting screenshots to result JSON #1056.
Adding screenshots in scripting is a great feature and one thing that
was missing was that the result JSON do not include any references to
the screenshots, so tools that uses browsertime, didn't know that they
exist. In sitespeed.ios case the screenshots are stored to disk but now shown.

This fix makes it possible to get the references from the JSON. In future
(next major) we should rename the screenshot files, they are now named:
name + "-index"

But it would have been more sane to have it the other way around (index first)
but don't want to potentially break any functionality so let us keep it for
now.