Fix missing SpeedIndex in HAR
Commit 206afb39332c2b0d10ed209568b610b9784ced0d introduced a bug where
the SpeedIndex and PerceptualSpeedIndex metrics are no longer added to
the HAR under "har.log.pages[]._visualMetrics". This is causing problems
with Sitespeed.io Compare, since it now can't find the SpeedIndex.

I'm not 100% sure why that commit changed the condition to
"key.indexOf('Progress') === 0)", since there aren't any metrics that
start with "Progress". I'm guessing the intent was to check if the
metric name ends with "Progress", so I changed it do that. I also added
tests.