Navigate to the same URL should generate multiple tests.
In 10 testing the same page twice in a script was broken as reported
in #1419 and also reported in https://github.com/sitespeedio/sitespeed.io/issues/3164.

The problem was that when we stop the video, we didn't know the correct URL,
that is fixed with this commit (+code cleanup to make it easier to follow).