Provide active tab id to gecko profiler StartProfiler API
In the `Services.profiler.StartProfiler` API we have an optional
parameter to pass the active tab id. This needs to be retrieved and
provided to the API so profiler can determine which tab was active at
the time of a profiling session.

This is optional because it's not a crucial profiler feature, but it's
useful to provide this.
You can see the code similar to this inside Firefox:
https://searchfox.org/mozilla-central/rev/d7d2cc647772de15c4c5aa47f74d25d0e379e404/devtools/client/performance-new/popup/background.jsm.js#382-393
and
https://searchfox.org/mozilla-central/rev/d7d2cc647772de15c4c5aa47f74d25d0e379e404/devtools/shared/performance-new/recording-utils.js#16-34

While I was changing this part, I also added some more context in the
commments for the API changes that we've done in the past. Hope it's
useful!