Add power consumption of all the processes while calculating the total power consumption
Gecko profile json includes child processes inside it's `profile.processes`
array. It's recursive data type that has the same properties. Previously
we were only calculating the power consumption of the parent process.
With this change, we should be able to add the consumption of all the
processes, including the content process (which has the most impact),
utility processes, webextension process etc.