Do not add the individual cpu component power values since 'CPU Package' will include them
For Linux and Windows power values there is always a counter called 'CPU Package',
this is the one we usually want. And besides this one, we might have
individual CPU components, like DRAM, iGPU, individual CPU cores etc.

These values are coming from individual components, but they are already
included inside the 'CPU Package' metrics. So this means that in these
platforms, if these individual components are captured, we are counting
the power usage values twice. The values could look higher than it
should.

See the available tracks here:
Linux:
https://searchfox.org/mozilla-central/rev/5ad94327fc03d0e5bc6aa81c0d7d113c2dccf947/tools/profiler/core/PowerCounters-linux.cpp#48-70
Windows:
https://searchfox.org/mozilla-central/rev/5ad94327fc03d0e5bc6aa81c0d7d113c2dccf947/tools/profiler/core/PowerCounters-win.cpp#32-55

So we shouldn't add them on top of the whole CPU package number,
otherwise it might look like we consume double the power that CPU
normally consumes.