Ignore the first counter value as it might be incorrect or unreliable
This is already done in the profiler frontend here:
https://github.com/firefox-devtools/profiler/blob/6a683fa6ae2733dae49ff460edf6ee98ca626f09/src/profile-logic/profile-data.js#L1740-L1743

Some platforms sometimes provide incorrect or unrealistic values. We
should remove these values so it doesn't generate an incorrect value.
Since this is only one value in a profile, we are not really concerned
in case it's a correct number. It's a very small difference in a big
array, removing this value is always safer.