Add Simpleperf app_profiler options and use --android.simpleperf.
Extend commands.simpleperf.start() to support for Simpleperf's general
app_profiler options. This allows users to pass options like
Simpleperf's -nb flag which skips generating a binary cache i.e. a
dumped symbols directory (this is useful when symbols are not needed
after profiling or when Simpleperf is failing to create a binary cache).

Replace --android.ndk with --android.simpleperf. --android.simpleperf is
more flexible as it expects the path directly to the directory
containing Simpleperf (whereas --android.ndk expects the directory to
the NDK installation, which is then used to find Simpleperf via a
hardcoded relative path: simpleperf/app_profiler.py). This is
particularly useful when Simpleperf is provided standalone, without the
NDK (as is the case with a Simpleperf repack).